moves the select form to it's own component
